Refactor checkout page and move into sendtag route
